Wires up account profile

+++ lennys-bakery/src/main/java/com/loco/service/UserService.java
@@ -21,7 +21,7 @@ public class UserService {
      *
      * @return int userId
      */
-    public int getUserIdFromSession() {
+    public long getUserIdFromSession() {
         return 2;
     }
 }
